- Moved TrainJobForm.js and PredJobForm.js
