+++ src/Console/Job Forms/PredJobForm.js
@@ -1,20 +1,10 @@
 import React, {useState} from "react";
-import {
-  Button, CircularProgress,
-  Dialog,
-  DialogTitle,
-  Hidden,
-  makeStyles,
-  MenuItem,
-  Slide,
-  TextField
-} from "@material-ui/core";
-import {submitJob} from "../common/Managers/EndpointManager";
-import {useHistory} from "react-router-dom";
-import {useAuth} from "../common/Auth.js";
-import AWSImportView from "./AWSImport/AWSImportView";
+import {Button, Dialog, DialogTitle, Hidden, makeStyles, MenuItem, Slide, TextField} from "@material-ui/core";
+import {useAuth} from "../../common/Auth.js";
+import AWSImportView from "../AWSImport/AWSImportView";
 import DoneIcon from '@material-ui/icons/Done';
 import ClearIcon from '@material-ui/icons/Clear';
+import {useHistory} from "react-router-dom";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -37,30 +27,18 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-// Useful constants
-const minJobTimeValue = 5;
-const maxJobTimeValue = 2880; // 48 hours
-const jobTimeOptions = [
-  {name: "minutes", value: 1},
-  {name: "hours", value: 60}
-];
-
 const SlideUpTransition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
 });
 
-
-export default function TrainJobForm() {
+/**
+ * @param props.header should contain the header of the train job
+ */
+export default function JobForm(props) {
   const [jobName, setJobName] = useState("");
   const [maxJobTime, setMaxJobTime] = useState(10);
   const [timeOption, setTimeOption] = useState(1);
-  const [targetColumn, setTargetColumn] = useState("");
-  const [header, setHeader] = useState([]);
   const [CSV, setCSV] = useState("");
-
-  const [isLoadingFile, setIsLoadingFile] = useState(false);
-  const [loadingValue, setLoadingValue] = useState(0);
-  const [progressBarType, setProgressBarType] = useState('determinate');
 
   const [showModal, setShowModal] = useState(false);
   const [dataImportSuccess, setDataImportSuccess] = useState(undefined);
@@ -69,6 +47,15 @@ export default function TrainJobForm() {
   const auth = useAuth();
   const history = useHistory();
   const fileInput = React.createRef();
+
+  // Useful constants
+  const minJobTimeValue = 5;
+  const maxJobTimeValue = 2880; // 48 hours
+  const jobTimeOptions = [
+    {name: "minutes", value: 1},
+    {name: "hours", value: 60}
+  ];
+  const jobTime = maxJobTime * timeOption;
 
   // Functions
   function handleMaxJobTimeChange(event) {
@@ -89,13 +76,16 @@ export default function TrainJobForm() {
       const reader = new FileReader();
       reader.readAsText(file);
       reader.onload = () => resolve(reader.result);
-      reader.onprogress = (progress) => {
-        // Note: We only update if we got to the next percentage point. Otherwise, too many state updates slow down file loading
-        const newProgressValue = Math.round(progress.loaded/progress.total * 100.);
-        if (newProgressValue !== loadingValue) setLoadingValue(newProgressValue);
-      };
       reader.onerror = (err) => reject(err);
     });
+  }
+
+  function isEqualArrays(arr1, arr2) {
+    if (arr1.length !== arr2.length) return false;
+    for (let i = 0; i < arr1.length; i++) {
+      if (arr1[i] !== arr2[i]) return false;
+    }
+    return true;
   }
 
   // converts array of fields into array of json objects
@@ -103,47 +93,37 @@ export default function TrainJobForm() {
     setCSV(csvString);
     const header = csvString.split('\n')[0];
     const fields = header.split(',');
-    console.log(fields);
-    setHeader(fields.map((field, ind) => {
-      return {name: field, col: ind};
-    }));
-    setTargetColumn(fields[fields.length - 1]); // Default to last column, as usually the last one is the target col
-    setDataImportSuccess(true);
-  }
-
-  function onFilePicked() {
-    console.log("file picked");
-    const file = fileInput.current.files[0];
-    if (file.name.substring(file.name.length - 4) !== '.csv') alert("File name must have a .csv extension");
-    else {
-      setDataImportSuccess(undefined); // Set both so success/fail messages go away
-      setIsLoadingFile(true);
-      setProgressBarType('determinate');
-      getFileObjectContent(file)
-        .then(csvString => {
-          updateCSVState(csvString);
-        })
-        .catch(err => {
-          console.log(err);
-          setDataImportSuccess(false);
-        })
-        .finally(() => setIsLoadingFile(false));
+    if (!isEqualArrays(fields, props.header)) {
+      alert("The prediction dataset must have the same columns as the training dataset");
+      setDataImportSuccess(false);
+    } else {
+      setDataImportSuccess(true);
     }
   }
 
-  function validateFormData(jobTime) {
+  function onFilePicked() {
+    const file = fileInput.current.files[0];
+    if (file.name.substring(file.name.length - 4) !== '.csv') alert("File name must have a .csv extension");
+    else {
+      getFileObjectContent(file)
+        .then(csvString => updateCSVState(csvString))
+        .catch(err => {
+          console.log(err);
+          setDataImportSuccess(false);
+        });
+    }
+  }
+
+  function validateFormData() {
     if (CSV === "") {
       alert("You must upload a csv file to train on.");
       return false;
-    // } else
+      // } else
     } else if (jobName.length === 0) {
       alert("Job name cannot be empty");
       return false;
     } else if (maxJobTime > maxJobTimeValue || jobTime < minJobTimeValue) {
       alert("Max Job Time must be between 10 minutes and 48 hours");
-      return false;
-    } else if (targetColumn.length === 0) {
-      alert("You must select a target column from the dropdown");
       return false;
     } else {
       return true;
@@ -152,11 +132,8 @@ export default function TrainJobForm() {
 
   function submitHandler(event) {
     event.preventDefault();
-    const jobTime = maxJobTime * timeOption;
     if (validateFormData()) {
-      submitJob(auth.user.email, jobName, jobTime, targetColumn, CSV)
-        .then(res => history.push('/console/jobs'))
-        .catch(err => alert("Job failed to submit"));
+      //TODO
     }
   }
 
@@ -164,10 +141,6 @@ export default function TrainJobForm() {
     <>
       <div className={classes.rootDiv}>
         <div>
-          <div hidden={!isLoadingFile}>
-            <p>Loading Dataset </p>
-            <CircularProgress variant={progressBarType} value={loadingValue} />
-          </div>
           <Hidden smUp={dataImportSuccess === undefined || !dataImportSuccess}>
             <DoneIcon color="primary"/>
             <p style={{color: "green"}}>Successfully imported</p>
@@ -179,25 +152,6 @@ export default function TrainJobForm() {
         </div>
 
         <form className={classes.root}>
-          <Hidden smUp={dataImportSuccess || isLoadingFile}>
-            {/*File Input Element*/}
-            <label htmlFor="fileInput">
-              <input type="file"
-                     accept=".csv"
-                     ref={fileInput}
-                     className={classes.fileInput}
-                     id="fileInput"
-                     name="File Input"
-                     onChange={onFilePicked}
-              />
-              <Button color="primary" variant="outlined" component="span">
-                Upload CSV
-              </Button>
-            </label>
-
-            {/*AWS Import button*/}
-            <Button variant="outlined" color="secondary" onClick={openModal}>Import from AWS S3</Button>
-          </Hidden>
 
           {/*Job Name*/}
           <div>
@@ -235,28 +189,25 @@ export default function TrainJobForm() {
             </TextField>
           </div>
 
-          {/*Target Column*/}
-          <div>
-            <Hidden smUp={!dataImportSuccess}>
-              <TextField
-                select
-                value={targetColumn}
-                id="columnName"
-                label="Target Column"
-                variant="outlined"
-                margin="normal"
-                onChange={(event) => setTargetColumn(event.target.value)}
-              >
-                {
-                  header.map((field) => (
-                    <MenuItem key={field.col} value={field.name}>
-                      {field.name}
-                    </MenuItem>
-                  ))
-                }
-              </TextField>
-            </Hidden>
-          </div>
+          <Hidden smUp={dataImportSuccess}>
+            {/*File Input Element*/}
+            <label htmlFor="fileInput">
+              <input type="file"
+                     accept=".csv"
+                     ref={fileInput}
+                     className={classes.fileInput}
+                     id="fileInput"
+                     name="File Input"
+                     onChange={onFilePicked}
+              />
+              <Button color="primary" variant="outlined" component="span">
+                Upload CSV
+              </Button>
+            </label>
+
+            {/*AWS Import button*/}
+            <Button variant="outlined" color="secondary" onClick={openModal}>Import from AWS S3</Button>
+          </Hidden>
 
           {/*Submit button*/}
           <div>
@@ -280,12 +231,10 @@ export default function TrainJobForm() {
         aria-labelledby="Import from AWS Form"
         aria-describedby="Input your AWS User details here"
         TransitionComponent={SlideUpTransition}
-        keepMounted
       >
         <DialogTitle>Import from AWS</DialogTitle>
         <AWSImportView closeModal={closeModal} setFile={setCSV} setDataImportSuccess={setDataImportSuccess}
-                       updateCSVState={updateCSVState} setIsLoadingFile={setIsLoadingFile} setLoadingValue={setLoadingValue}
-                       setProgressBarType={setProgressBarType}
+                       updateCSVState={updateCSVState}
         />
       </Dialog>
     </>
