+++ src/Console/Job Forms/TrainJobForm.js
@@ -1,10 +1,20 @@
 import React, {useState} from "react";
-import {Button, Dialog, DialogTitle, Hidden, makeStyles, MenuItem, Slide, TextField} from "@material-ui/core";
-import {useAuth} from "../common/Auth.js";
-import AWSImportView from "./AWSImport/AWSImportView";
+import {
+  Button, CircularProgress,
+  Dialog,
+  DialogTitle,
+  Hidden,
+  makeStyles,
+  MenuItem,
+  Slide,
+  TextField
+} from "@material-ui/core";
+import {submitJob} from "../../common/Managers/EndpointManager";
+import {useHistory} from "react-router-dom";
+import {useAuth} from "../../common/Auth.js";
+import AWSImportView from "../AWSImport/AWSImportView";
 import DoneIcon from '@material-ui/icons/Done';
 import ClearIcon from '@material-ui/icons/Clear';
-import {useHistory} from "react-router-dom";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -27,18 +37,30 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
+// Useful constants
+const minJobTimeValue = 5;
+const maxJobTimeValue = 2880; // 48 hours
+const jobTimeOptions = [
+  {name: "minutes", value: 1},
+  {name: "hours", value: 60}
+];
+
 const SlideUpTransition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
 });
 
-/**
- * @param props.header should contain the header of the train job
- */
-export default function JobForm(props) {
+
+export default function TrainJobForm() {
   const [jobName, setJobName] = useState("");
   const [maxJobTime, setMaxJobTime] = useState(10);
   const [timeOption, setTimeOption] = useState(1);
+  const [targetColumn, setTargetColumn] = useState("");
+  const [header, setHeader] = useState([]);
   const [CSV, setCSV] = useState("");
+
+  const [isLoadingFile, setIsLoadingFile] = useState(false);
+  const [loadingValue, setLoadingValue] = useState(0);
+  const [progressBarType, setProgressBarType] = useState('determinate');
 
   const [showModal, setShowModal] = useState(false);
   const [dataImportSuccess, setDataImportSuccess] = useState(undefined);
@@ -47,15 +69,6 @@ export default function JobForm(props) {
   const auth = useAuth();
   const history = useHistory();
   const fileInput = React.createRef();
-
-  // Useful constants
-  const minJobTimeValue = 5;
-  const maxJobTimeValue = 2880; // 48 hours
-  const jobTimeOptions = [
-    {name: "minutes", value: 1},
-    {name: "hours", value: 60}
-  ];
-  const jobTime = maxJobTime * timeOption;
 
   // Functions
   function handleMaxJobTimeChange(event) {
@@ -76,16 +89,13 @@ export default function JobForm(props) {
       const reader = new FileReader();
       reader.readAsText(file);
       reader.onload = () => resolve(reader.result);
+      reader.onprogress = (progress) => {
+        // Note: We only update if we got to the next percentage point. Otherwise, too many state updates slow down file loading
+        const newProgressValue = Math.round(progress.loaded/progress.total * 100.);
+        if (newProgressValue !== loadingValue) setLoadingValue(newProgressValue);
+      };
       reader.onerror = (err) => reject(err);
     });
-  }
-
-  function isEqualArrays(arr1, arr2) {
-    if (arr1.length !== arr2.length) return false;
-    for (let i = 0; i < arr1.length; i++) {
-      if (arr1[i] !== arr2[i]) return false;
-    }
-    return true;
   }
 
   // converts array of fields into array of json objects
@@ -93,37 +103,46 @@ export default function JobForm(props) {
     setCSV(csvString);
     const header = csvString.split('\n')[0];
     const fields = header.split(',');
-    if (!isEqualArrays(fields, props.header)) {
-      alert("The prediction dataset must have the same columns as the training dataset");
-      setDataImportSuccess(false);
-    } else {
-      setDataImportSuccess(true);
-    }
+    setHeader(fields.map((field, ind) => {
+      return {name: field, col: ind};
+    }));
+    setTargetColumn(fields[fields.length - 1]); // Default to last column, as usually the last one is the target col
+    setDataImportSuccess(true);
   }
 
   function onFilePicked() {
+    console.log("file picked");
     const file = fileInput.current.files[0];
     if (file.name.substring(file.name.length - 4) !== '.csv') alert("File name must have a .csv extension");
     else {
+      setDataImportSuccess(undefined); // Set both so success/fail messages go away
+      setIsLoadingFile(true);
+      setProgressBarType('determinate');
       getFileObjectContent(file)
-        .then(csvString => updateCSVState(csvString))
+        .then(csvString => {
+          updateCSVState(csvString);
+        })
         .catch(err => {
           console.log(err);
           setDataImportSuccess(false);
-        });
+        })
+        .finally(() => setIsLoadingFile(false));
     }
   }
 
-  function validateFormData() {
+  function validateFormData(jobTime) {
     if (CSV === "") {
       alert("You must upload a csv file to train on.");
       return false;
-      // } else
+    // } else
     } else if (jobName.length === 0) {
       alert("Job name cannot be empty");
       return false;
     } else if (maxJobTime > maxJobTimeValue || jobTime < minJobTimeValue) {
       alert("Max Job Time must be between 10 minutes and 48 hours");
+      return false;
+    } else if (targetColumn.length === 0) {
+      alert("You must select a target column from the dropdown");
       return false;
     } else {
       return true;
@@ -132,15 +151,29 @@ export default function JobForm(props) {
 
   function submitHandler(event) {
     event.preventDefault();
+    const jobTime = maxJobTime * timeOption;
     if (validateFormData()) {
-      //TODO
+      submitJob(auth.user.email, jobName, jobTime, targetColumn, CSV)
+        .then(res => history.push('/console/jobs'))
+        .catch(err => alert("Job failed to submit"));
     }
+  }
+
+  function _AWSImportView(props) {
+    return <AWSImportView setFile={setCSV} setDataImportSuccess={setDataImportSuccess}
+                          updateCSVState={updateCSVState} setIsLoadingFile={setIsLoadingFile}
+                          setLoadingValue={setLoadingValue}
+                          setProgressBarType={setProgressBarType} {...props}/>;
   }
 
   return (
     <>
       <div className={classes.rootDiv}>
         <div>
+          <div hidden={!isLoadingFile}>
+            <p>Loading Dataset </p>
+            <CircularProgress variant={progressBarType} value={loadingValue} />
+          </div>
           <Hidden smUp={dataImportSuccess === undefined || !dataImportSuccess}>
             <DoneIcon color="primary"/>
             <p style={{color: "green"}}>Successfully imported</p>
@@ -152,6 +185,25 @@ export default function JobForm(props) {
         </div>
 
         <form className={classes.root}>
+          <Hidden smUp={dataImportSuccess || isLoadingFile}>
+            {/*File Input Element*/}
+            <label htmlFor="fileInput">
+              <input type="file"
+                     accept=".csv"
+                     ref={fileInput}
+                     className={classes.fileInput}
+                     id="fileInput"
+                     name="File Input"
+                     onChange={onFilePicked}
+              />
+              <Button color="primary" variant="outlined" component="span">
+                Upload CSV
+              </Button>
+            </label>
+
+            {/*AWS Import button*/}
+            <Button variant="outlined" color="secondary" onClick={openModal}>Import from AWS S3</Button>
+          </Hidden>
 
           {/*Job Name*/}
           <div>
@@ -189,25 +241,28 @@ export default function JobForm(props) {
             </TextField>
           </div>
 
-          <Hidden smUp={dataImportSuccess}>
-            {/*File Input Element*/}
-            <label htmlFor="fileInput">
-              <input type="file"
-                     accept=".csv"
-                     ref={fileInput}
-                     className={classes.fileInput}
-                     id="fileInput"
-                     name="File Input"
-                     onChange={onFilePicked}
-              />
-              <Button color="primary" variant="outlined" component="span">
-                Upload CSV
-              </Button>
-            </label>
-
-            {/*AWS Import button*/}
-            <Button variant="outlined" color="secondary" onClick={openModal}>Import from AWS S3</Button>
-          </Hidden>
+          {/*Target Column*/}
+          <div>
+            <Hidden smUp={!dataImportSuccess}>
+              <TextField
+                select
+                value={targetColumn}
+                id="columnName"
+                label="Target Column"
+                variant="outlined"
+                margin="normal"
+                onChange={(event) => setTargetColumn(event.target.value)}
+              >
+                {
+                  header.map((field) => (
+                    <MenuItem key={field.col} value={field.name}>
+                      {field.name}
+                    </MenuItem>
+                  ))
+                }
+              </TextField>
+            </Hidden>
+          </div>
 
           {/*Submit button*/}
           <div>
@@ -231,10 +286,12 @@ export default function JobForm(props) {
         aria-labelledby="Import from AWS Form"
         aria-describedby="Input your AWS User details here"
         TransitionComponent={SlideUpTransition}
+        keepMounted
       >
         <DialogTitle>Import from AWS</DialogTitle>
         <AWSImportView closeModal={closeModal} setFile={setCSV} setDataImportSuccess={setDataImportSuccess}
-                       updateCSVState={updateCSVState}
+                       updateCSVState={updateCSVState} setIsLoadingFile={setIsLoadingFile} setLoadingValue={setLoadingValue}
+                       setProgressBarType={setProgressBarType}
         />
       </Dialog>
     </>
